requirements.txt
# Use psycopg2==2.9.10 only if option psycopg2-binary does not work
psycopg2-binary~=2.9.10
python-dotenv~=1.1.1
requests~=2.32.4
beautifulsoup4~=4.13.4
cryptography~=45.0.6
fastapi~=0.111.1
uvicorn~=0.30.3